[bluelink] Allow float values for HV & 12V battery SoC

+++ bundles/org.openhab.binding.bluelink/src/main/java/org/openhab/binding/bluelink/internal/dto/EvStatus.java
@@ -21,7 +21,7 @@ import com.google.gson.annotations.SerializedName;
  *
  * @author Marcus Better - Initial contribution
  */
-public record EvStatus(boolean batteryCharge, int batteryStatus, @SerializedName("batteryPlugin") int rawBatteryPlugin,
+public record EvStatus(boolean batteryCharge, double batteryStatus, @SerializedName("batteryPlugin") int rawBatteryPlugin,
         ReserveChargeInfo reservChargeInfos, List<DrivingDistance> drvDistance, ChargeRemainingTime remainTime2) {
 
     public boolean batteryPlugin() {
